Undoing error handling around `calcAndCheckSum`
due to errors being ignored in code as a way to step past missing
checksum files when you want to create them, normal error validation
caused issues in the normal execution of the restore process.
removing checks until future work to resolve these patterns in the
code can be taken on.